fix 5

+++ libft/ft_strlcpy.c
@@ -6,13 +6,13 @@
 /*   By: elaachac <elaachac@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/05 16:51:47 by elaachac          #+#    #+#             */
-/*   Updated: 2019/11/18 15:29:42 by elaachac         ###   ########.fr       */
+/*   Updated: 2019/11/18 18:23:50 by elaachac         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-unsigned int	ft_strlcpy(char *dest, const char *src, size_t size)
+size_t	ft_strlcpy(char *dest, const char *src, size_t size)
 {
 	unsigned int i;
 	unsigned int j;
